created new file file TokenSwitch for switching functionality and solved problem with NaN when no value was entered
